Update and rename login_page.dart to login_page+register_page+blank_page.dart
