+++ login_page+register_page+blank_page.dart
@@ -1,0 +1,244 @@
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      home: LoginDemo(),
+    );
+  }
+}
+
+class LoginDemo extends StatefulWidget {
+  const LoginDemo({Key? key}) : super(key: key);
+
+  @override
+  LoginDemoState createState() => LoginDemoState();
+}
+
+// --------- Login Page ---------------//
+class LoginDemoState extends State<LoginDemo> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      //appBar: AppBar(title: Text('Login'),),
+      //this is the parent of body
+      body: Container(
+        child: Column(
+          children: <Widget>[
+            Padding(
+              child: Text('Login',
+                  style: TextStyle(
+                    color: Colors.blue,
+                    fontSize: 30,
+                  )),
+              padding: const EdgeInsets.only(top: 30.0),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 5.0),
+              child: Center(
+                child: Container(
+                    width: 150,
+                    height: 150,
+                    /*decoration: BoxDecoration(
+                        color: Colors.red,
+                        borderRadius: BorderRadius.circular(50.0)),*/
+                    //thanks to frrepik for the resource
+                    child: Image.network(
+                        'https://img.freepik.com/free-vector/people-putting-puzzle-pieces-together_52683-28610.jpg?size=626&ext=jpg')),
+              ),
+            ),
+            Padding(
+              //padding: const EdgeInsets.only(left:15.0,right: 15.0,top:0,bottom: 0),
+              padding: EdgeInsets.symmetric(horizontal: 15),
+              child: TextField(
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Email',
+                    hintText: 'Enter valid email id as abc@gmail.com'),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(
+                  left: 15.0, right: 15.0, top: 15, bottom: 0),
+              //padding: EdgeInsets.symmetric(horizontal: 15),
+              child: TextField(
+                obscureText: true,
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Password',
+                    hintText: 'Enter secure password'),
+              ),
+            ),
+            FlatButton(
+              onPressed: () {
+                //TODO FORGOT PASSWORD SCREEN GOES HERE
+              },
+              child: Text(
+                'Forgot Password',
+                style: TextStyle(color: Colors.blue, fontSize: 15),
+              ),
+            ),
+            Container(
+              height: 50,
+              width: 250,
+              decoration: BoxDecoration(
+                  color: Colors.blue, borderRadius: BorderRadius.circular(20)),
+              child: TextButton(
+                onPressed: () {
+                  Navigator.push(
+                      context, MaterialPageRoute(builder: (_) => IsiChat()));
+                },
+                child: Text(
+                  'Login',
+                  style: TextStyle(color: Colors.white, fontSize: 25),
+                ),
+              ),
+            ),
+            SizedBox(
+              height: 30,
+              width: 10,
+            ),
+
+            TextButton(child: Text('Register here'), onPressed: () {
+              Navigator.push(
+                      context, MaterialPageRoute(builder: (_) => HalamanDaftar()));
+            }),
+            //Text('New User? Create Account')
+          ],
+          
+        ),
+      ),
+    );
+  }
+}
+
+// -----------tampilan chat------------ //
+
+class IsiChat extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        title: Text('Isi Chat'),
+      ),
+      //this is the parent of body
+      body: Container(
+        child: Column(
+          //place the content here(listview.builder)
+          children: [],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.pop(context);
+        },
+      ),
+    );
+  }
+}
+
+//-----------tampilan register--------------//
+
+class HalamanDaftar extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      //appBar: AppBar(title: Text('Isi Chat'),),
+      //this is the parent of body
+      body: Container(
+        child: Column(
+          children: <Widget>[
+            Padding(
+              child: Text('Register',
+                  style: TextStyle(
+                    color: Colors.blue,
+                    fontSize: 30,
+                  )),
+              padding: const EdgeInsets.only(top: 30.0),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(top: 5.0),
+              child: Center(
+                child: Container(
+                  
+                    width: 150,
+                    height: 150,
+                    /*decoration: BoxDecoration(
+                        color: Colors.red,
+                        borderRadius: BorderRadius.circular(50.0)),*/
+                    //thanks to frrepik for the resource
+                    child: Image.network(
+                        'https://img.freepik.com/free-vector/people-putting-puzzle-pieces-together_52683-28610.jpg?size=626&ext=jpg')),
+              ),
+            ),
+            Padding(
+              //padding: const EdgeInsets.only(left:15.0,right: 15.0,top:0,bottom: 0),
+              padding: EdgeInsets.symmetric(horizontal: 15),
+              child: TextField(
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Email',
+                    hintText: 'Enter valid email id as abc@gmail.com'),
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.only(
+                  left: 15.0, right: 15.0, top: 15, bottom: 0),
+              //padding: EdgeInsets.symmetric(horizontal: 15),
+              child: TextField(
+                obscureText: true,
+                decoration: InputDecoration(
+                    border: OutlineInputBorder(),
+                    labelText: 'Password',
+                    hintText: 'Enter secure password'),
+              ),
+            ),
+            FlatButton(
+              onPressed: () {
+                //TODO FORGOT PASSWORD SCREEN GOES HERE
+              },
+              child: Text(
+                'Forgot Password',
+                style: TextStyle(color: Colors.blue, fontSize: 15),
+              ),
+            ),
+            Container(
+              height: 50,
+              width: 250,
+              decoration: BoxDecoration(
+                  color: Colors.blue, borderRadius: BorderRadius.circular(20)),
+              child: FlatButton(
+                onPressed: () {
+                  Navigator.push(
+                      context, MaterialPageRoute(builder: (_) => IsiChat()));
+                },
+                child: Text(
+                  'Login',
+                  style: TextStyle(color: Colors.white, fontSize: 25),
+                ),
+              ),
+            ),
+            SizedBox(
+              height: 130,
+            ),
+            //Text('New User? Create Account')
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          Navigator.pop(context);
+        },
+      ),
+    );
+  }
+}
